+++ docs_src/migrations/accounts_models.py
@@ -4,7 +4,7 @@ from my_project.utils import get_db_connection
 
 import edgy
 
-_, registry = get_db_connection()
+registry = get_db_connection()
 
 
 class User(edgy.Model):