+++ docs_src/migrations/attaching.py
@@ -1,46 +1,10 @@
-#!/usr/bin/env python
-import os
-import sys
-from pathlib import Path
+from typing import Optional, Union
 
-from my_project.utils import get_db_connection
-
-from edgy import Instance, monkay
-from esmerald import Esmerald, Include
+from edgy import EdgySettings
 
 
-def build_path():
-    """
-    Builds the path of the project and project root.
-    """
-    Path(__file__).resolve().parent.parent
-    SITE_ROOT = os.path.dirname(os.path.realpath(__file__))
-
-    if SITE_ROOT not in sys.path:
-        sys.path.append(SITE_ROOT)
-        sys.path.append(os.path.join(SITE_ROOT, "apps"))
-
-
-def get_application():
-    """
-    This is optional. The function is only used for organisation purposes.
-    """
-    build_path()
-    registry = get_db_connection()
-
-    app = registry.asgi(
-        Esmerald(
-            routes=[Include(namespace="my_project.urls")],
-        )
-    )
-
-    monkay.set_instance(
-        Instance(
-            app=app,
-            registry=registry,
-        )
-    )
-    return app
-
-
-app = get_application()
+class MyMigrationSettings(EdgySettings):
+    # here we notify about the models import path
+    preloads: list[str] = ["myproject.apps.accounts.models"]
+    # here we can set the databases which should be used in migrations, by default (None,)
+    migrate_databases: Union[list[Union[str, None]], tuple[Union[str, None], ...]] = (None,)