+++ docs_src/migrations/lru.py
@@ -5,5 +5,6 @@ from edgy import Database, Registry
 
 @lru_cache()
 def get_db_connection():
-    database = Database("postgresql+asyncpg://user:pass@localhost:5432/my_database")
-    return database, Registry(database=database)
+    # use echo=True for getting the connection infos printed
+    database = Database("postgresql+asyncpg://user:pass@localhost:5432/my_database", echo=True)
+    return Registry(database=database)