+++ edgy/cli/templates/url/env.py
@@ -27,13 +27,14 @@ config: Any = context.config
 # This line sets up loggers basically.
 fileConfig(config.config_file_name)
 logger = logging.getLogger("alembic.env")
+MAIN_DATABASE_NAME: str = " "
 
 
 def iter_databases(
     registry: Registry,
 ) -> Generator[tuple[str, Database, "sqlalchemy.MetaData"], None, None]:
     url: Optional[str] = os.environ.get("EDGY_DATABASE_URL")
-    name: Union[str, Literal[False]] = os.environ.get("EDGY_DATABASE") or False
+    name: Union[str, Literal[False], None] = os.environ.get("EDGY_DATABASE") or False
     if url and not name:
         try:
             name = registry.metadata_by_url.get_name(url)
@@ -47,6 +48,8 @@ def iter_databases(
             else:
                 yield (name, registry.extra[name], registry.metadata_by_name[name])
     else:
+        if name == MAIN_DATABASE_NAME:
+            name = None
         if url:
             database = Database(url)
         elif name is None: